requirements.txt
dnspython==1.16.0
pymongo==3.11.4
pyserial==3.5
python-dotenv==0.17.1
